fix crash when no pump settings
