+++ packages/yourloops/pages/device/device-page.tsx
@@ -28,91 +28,37 @@
 import React, { type FC } from 'react'
 import type MedicalDataService from 'medical-domain'
 import Container from '@mui/material/Container'
-import Card from '@mui/material/Card'
-import CardContent from '@mui/material/CardContent'
-import CardHeader from '@mui/material/CardHeader'
 import { useTranslation } from 'react-i18next'
-import Button from '@mui/material/Button'
-import FileCopyIcon from '@mui/icons-material/FileCopy'
 import { useTheme } from '@mui/material/styles'
-import { useDevice } from '../../components/device/use-device.hook'
-import Grid from '@mui/material/Grid'
-import { makeStyles } from 'tss-react/mui'
-import { DeviceInfo } from '../../components/device/device-info'
-import { PumpInfo } from '../../components/device/pump-info'
-import { CgmInfo } from '../../components/device/cgm-info'
-import { ParameterList } from '../../components/device/parameter-list'
-import { ParametersChangeHistory } from '../../components/device/parameters-change-history'
 import Box from '@mui/material/Box'
+import { DeviceSettings } from '../../components/device/device-settings'
+import Typography from '@mui/material/Typography'
 
 interface DevicePageProps {
   goToDailySpecificDate: (date: number) => void
   medicalData: MedicalDataService
 }
 
-const useStyles = makeStyles()(() => ({
-  cardHeaderAction: {
-    marginTop: 0
-  }
-}))
-
 export const DevicePage: FC<DevicePageProps> = ({ medicalData, goToDailySpecificDate }) => {
-  const theme = useTheme()
-  const { classes } = useStyles()
   const { t } = useTranslation()
-  const {
-    copySettingsToClipboard,
-    lastUploadDate,
-    device,
-    pump,
-    parameters,
-    history,
-    cgm
-  } = useDevice(medicalData)
+  const theme = useTheme()
 
   return (
     <Container data-testid="device-settings-container">
-      <Card variant="outlined" sx={{ padding: theme.spacing(2) }}>
-        <CardHeader
-          title={t('device')}
-          subheader={`${t('last-upload:')} ${lastUploadDate}`}
-          action={
-            <Button
-              variant="contained"
-              disableElevation
-              startIcon={<FileCopyIcon />}
-              onClick={copySettingsToClipboard}
-            >
-              {t('text-copy')}
-            </Button>
-          }
-          classes={{
-            action: classes.cardHeaderAction
-          }}
+      {medicalData.hasDeviceSettings()
+        ? <DeviceSettings
+          goToDailySpecificDate={goToDailySpecificDate}
+          medicalData={medicalData}
         />
-        <CardContent>
-          <Grid
-            container
-            spacing={4}
-            rowSpacing={4}
-          >
-            <Grid item xs={12} sm={6}>
-              <DeviceInfo device={device} />
-              <PumpInfo pump={pump} />
-              <CgmInfo cgm={cgm} />
-            </Grid>
-            <Grid item xs={12} sm={6} data-testid="parameters-container">
-              <ParameterList parameters={parameters} />
-            </Grid>
-          </Grid>
-          <Box marginTop={5}>
-            <ParametersChangeHistory
-              onClickChangeDate={goToDailySpecificDate}
-              history={history}
-            />
-          </Box>
-        </CardContent>
-      </Card>
+        : <Box
+          display="flex"
+          justifyContent="center"
+          alignItems="center"
+          marginTop={theme.spacing(4)}
+        >
+          <Typography fontWeight={500}>{t('no-settings-on-device-alert-message')}</Typography>
+        </Box>
+      }
     </Container>
   )
 }
